adding index.md
this is a clever commit title given that it both adds an index.md file as a demo, but it also solved an issue where if you're using the web extension to view a markus document by just localhost:88 it wouldn't be a url ending in .md so your web extension wouldn't know to parse it as markdown. so this way now if you pull up localhost:88 you end up on localhost:88/index.md. oh also and theres a 404 message now.